Moved gazebo logic in a new GazeboWrapper class
It will be used to create gym environments directly from python.
