+++ bindings/CMakeLists.txt
@@ -16,7 +16,7 @@ set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
 
 set_property(SOURCE gympp.i PROPERTY CPLUSPLUS ON)
 swig_add_library(gympp LANGUAGE python SOURCES gympp.i)
-swig_link_libraries(gympp GymFactory IgnitionEnvironment RobotSingleton)
+swig_link_libraries(gympp GymFactory IgnitionEnvironment RobotSingleton GazeboWrapper)
 
 set(target ${SWIG_MODULE_gympp_REAL_NAME})
 #add_dependencies(${target} ${CMAKE_CURRENT_SOURCE_DIR}/optional.i)
